Update SwerveModuleSDS.java
Added some PID values to steering because it was not turning probably because Proportional power p was set to zero so no power was ever provided.  Just ysing 0.001 that worked last year it seems, at least that was in the code, make some sense because error will be in Degrees so if 90 degrees of that would set the power to 9% power, and being not much power is needed to turn the wheel that 9% is probably enough.