SwerveDrive Config IMprovments
Fix turn gear ratio & now pass in max speeed to drive. seems to drive real good (according to coach Rich), only problem is Path Planner's rotations are the opposite diretions and that throws the paths off. (so that still needs to be fixxed) but storing away this stable point. Oh and right now Robot Drive relative for joystick but can be switched back to Field oriented.