NegatedControlSoTeleop Works & NewElevetorHeights
On Thursdy Before troy made improvments for PathPLanner, (moved negating out of drive) and that made pathplanner better. but that would mean for Seth it would not drive the same so put negating in the Joystick control.

Also recalculated the Eleveator to allow the bottom to be negative (below the starting posiotn that has the arm down)